Fix MSVC C4244/C4267 narrowing conversion warnings
- send_game_select: return int instead of int8_t to match send_all_tcp
- net.c: explicit casts at the protobuf boundary where protobuf-c forces
  int32_t/size_t but the wire protocol and app types are narrower

+++ src/game.c
@@ -122,7 +122,7 @@ const GameChoice_t *find_game_choice_by_type(const uint8_t type) {
   return NULL; // Not found
 }
 
-int8_t send_game_select(TCPsocket sock, uint8_t game_type, bool deuces_wild) {
+int send_game_select(TCPsocket sock, uint8_t game_type, bool deuces_wild) {
   GameSelectPayload_t payload = {game_type, deuces_wild ? 1 : 0};
 
   const uint32_t payload_size = OPCODE_SIZE + sizeof(payload);
